+++ test/common-test/pipeline-test.cpp
@@ -5,43 +5,19 @@
 #include <vector>
 #include <memory>
 
-#include "src/common/pipeline/pipelines.hpp"
 #include "src/common/style.hpp"
 #include "src/distance/edge.hpp"
 
 #include "test/common/constants/pipeline-constants.hpp"
 #include "test/common/mocks/pipeline-mocks.hpp"
+#include "test/common/mocks/distance-mocks.hpp"
+#include "test/common/common.hpp"
+
+#include "src/common/pipeline/pipelines.hpp"
 
 namespace found {
 
-// Lightweight concrete stages for direct testing without mocks.
-class ImageToPointsStage : public FunctionStage<Image, Points> {
- public:
-    Points Run(const Image &input) override {
-        (void)input;
-        return {{1, 2}, {3, 4}};
-    }
-};
-
-class PointsToVec3Stage : public FunctionStage<Points, Vec3> {
- public:
-    Vec3 Run(const Points &input) override {
-        (void)input;
-        return {5, 6, 7};
-    }
-};
-
-class IdentityPointsStage : public FunctionStage<Points, Points> {
- public:
-    Points Run(const Points &input) override {
-        Points output = input;
-        for (auto &pt : output) {
-            pt.x += DECIMAL(1.0);
-            pt.y += DECIMAL(2.0);
-        }
-        return output;
-    }
-};
+using Orientations = std::pair<EulerAngles, EulerAngles>;
 
 /////////////////////////////////////////////
 ///////// SEQUENTIAL PIPELINE TEST //////////
@@ -662,89 +638,132 @@ TEST(ModifyingPipelineTest, TestSequentialAndModifierPipelinesinSequentialPipeli
     ASSERT_EQ(expectedResource, resource);
 }
 
-// Alias-based smoke test for DistancePipeline to ensure the typedef stays wired
-TEST(SequentialPipelineTest, RunDistancePipelineAlias) {
-    DistancePipeline pipeline;
-    ImageToPointsStage stage1;
-    PointsToVec3Stage stage2;
+////////////////////////////////////////////
+////////// NOMINAL PIPELINE TEST ///////////
+////////////////////////////////////////////
 
-    Image input{};
-    PositionVector result = pipeline.AddStage(stage1).Complete(stage2).Run(input);
+TEST(NominalPipelineTest, TestPipelinesAsStages) {
+    // skip calibration stage, because it only has 1 stage
 
-    ASSERT_EQ(result.x, 5);
-    ASSERT_EQ(result.y, 6);
-    ASSERT_EQ(result.z, 7);
+    DistancePipeline distancePipeline;
+    PositionVector expectedVec(1, 2, 3);
+
+    // Here, we make a single stage that looks like the distance pipeline
+    std::unique_ptr<MockFunctionStage<Image, PositionVector>>
+        distanceStage(new MockFunctionStage<Image, PositionVector>());
+    EXPECT_CALL(*distanceStage, Run(testing::_))
+        .WillOnce(testing::Return(expectedVec));
+
+    PositionVector actualVec(distancePipeline.Complete(*distanceStage)
+                                             .Run({}));
+
+    ASSERT_VEC3_EQ_DEFAULT(expectedVec, actualVec);
+
+    OrbitPipeline orbitPipeline;
+    LocationRecords expectedLR;
+
+    // Again, we make a single stage regardless of how many there
+    // actually are in the pipeline
+    std::unique_ptr<MockFunctionStage<LocationRecords, LocationRecords>>
+        orbitStage(new MockFunctionStage<LocationRecords, LocationRecords>());
+    EXPECT_CALL(*orbitStage, Run(testing::_))
+        .WillOnce(testing::Return(expectedLR));
+
+    orbitPipeline.Complete(*orbitStage);
+
+    OrbitPipeline wrapperOrbitPipeline;
+    wrapperOrbitPipeline.Complete(orbitPipeline);
+    LocationRecords actualLR(wrapperOrbitPipeline.Run({}));
+
+    ASSERT_THAT(expectedLR, LocationRecordsEqual(actualLR));
 }
 
-TEST(SequentialPipelineTest, DoActionImageToVec3N10) {
-    SequentialPipeline<Image, Vec3, 10> pipeline;
-    ImageToPointsStage stage1;
-    PointsToVec3Stage stage2;
+TEST(NominalPipelineTest, TestNominalPipelinesWrapped) {
+    CalibrationPipeline calibrationPipeline;
+    Quaternion expectedQuat{1, 2, 3, 4};
 
-    pipeline.AddStage(stage1).Complete(stage2);
+    std::unique_ptr<MockFunctionStage<Orientations, Quaternion>>
+        calibrationStage(new MockFunctionStage<Orientations, Quaternion>());
+    EXPECT_CALL(*calibrationStage, Run(testing::_))
+        .WillOnce(testing::Return(expectedQuat));
 
-    Image input{};
-    Vec3 output{};
-    pipeline.GetResource() = input;
-    pipeline.GetProduct() = &output;
+    calibrationPipeline.Complete(*calibrationStage);
 
-    pipeline.DoAction(); // This directly calls the target function.
+    CalibrationPipeline wrapperCalibrationPipeline;
+    wrapperCalibrationPipeline.Complete(calibrationPipeline);
+    Quaternion actualQuat(wrapperCalibrationPipeline.Run({}));
 
-    ASSERT_EQ(output.x, 5);
-    ASSERT_EQ(output.y, 6);
-    ASSERT_EQ(output.z, 7);
+    ASSERT_QUAT_EQ_DEFAULT(expectedQuat, actualQuat);
+
+    DistancePipeline distancePipeline;
+    PositionVector expectedVec(1, 2, 3);
+
+    std::unique_ptr<MockEdgeDetectionAlgorithm>
+        edgeDetectionStage(new MockEdgeDetectionAlgorithm());
+    EXPECT_CALL(*edgeDetectionStage, Run(testing::_))
+        .WillOnce(testing::Return(Points()));
+
+    std::unique_ptr<MockDistanceDeterminationAlgorithm>
+        distanceStage(new MockDistanceDeterminationAlgorithm());
+    EXPECT_CALL(*distanceStage, Run(testing::_))
+        .WillOnce(testing::Return(PositionVector()));
+
+    std::unique_ptr<MockVectorGenerationAlgorithm>
+        vectorStage(new MockVectorGenerationAlgorithm());
+    EXPECT_CALL(*vectorStage, Run(testing::_))
+        .WillOnce(testing::Return(expectedVec));
+
+    distancePipeline.AddStage(*edgeDetectionStage)
+                    .AddStage(*distanceStage)
+                    .Complete(*vectorStage);
+
+    DistancePipeline wrapperDistancePipeline;
+    wrapperDistancePipeline.Complete(distancePipeline);
+    PositionVector actualVec(wrapperDistancePipeline.Run({}));
+
+    ASSERT_VEC3_EQ_DEFAULT(expectedVec, actualVec);
+
+    OrbitPipeline orbitPipeline;
+    LocationRecords expectedLR;
+
+    std::unique_ptr<MockFunctionStage<LocationRecords, LocationRecords>>
+        undefinedOrbitStage(new MockFunctionStage<LocationRecords, LocationRecords>());
+    EXPECT_CALL(*undefinedOrbitStage, Run(testing::_))
+        .WillOnce(testing::Return(expectedLR));
+
+    orbitPipeline.Complete(*undefinedOrbitStage);
+
+    OrbitPipeline wrapperOrbitPipeline;
+    wrapperOrbitPipeline.Complete(orbitPipeline);
+    LocationRecords actualLR(wrapperOrbitPipeline.Run({}));
+
+    ASSERT_THAT(expectedLR, LocationRecordsEqual(actualLR));
 }
 
-// Directly exercise FunctionStage<Points, Points>::DoAction
-TEST(FunctionStageTest, DoActionPointsIdentityStage) {
-    IdentityPointsStage stage;
-    Points input = {{DECIMAL(1.0), DECIMAL(2.0)}, {DECIMAL(3.0), DECIMAL(4.0)}};
-    Points product;
+TEST(NominalPipelineTest, TestNominalPipelinesModifyingPipelinesAsStages) {
+    // First, get all modifying stages as normal function stages
+    std::unique_ptr<MockFunctionStage<Points, Points>> edgeFilteringStage(
+        new MockFunctionStage<Points, Points>());
+    EXPECT_CALL(*edgeFilteringStage, Run(testing::_))
+        .WillOnce(testing::Return(points));
 
-    stage.GetResource() = input;
-    stage.GetProduct() = &product;
+    // Second, put them into a pipeline
+    INIT_SQ_PIPELINE(Points, Points, edgeFilteringPipeline);
+    edgeFilteringPipeline.Complete(*edgeFilteringStage);
 
-    stage.DoAction();
+    // Third, call and verify the output
+    Points actual(edgeFilteringPipeline.Run(points));
 
-    ASSERT_EQ(product.size(), input.size());
-    for (size_t i = 0; i < product.size(); ++i) {
-        EXPECT_EQ(product[i].x, input[i].x + DECIMAL(1.0));
-        EXPECT_EQ(product[i].y, input[i].y + DECIMAL(2.0));
-    }
-}
+    // What's happening here is that we use are getting all our elements
+    // in our expected output, wrapping them in a Matcher, and then
+    // asserting that the resulting array has the same elements in any order
+    std::vector<testing::Matcher<Vec2>> matchers;
+    std::transform(points.begin(), points.end(), std::back_inserter(matchers),
+                [](const Vec2& val) {
+                    return Vec2Equal(val);
+                });
 
-// Explicit coverage for SequentialPipeline<Image, Vec3, 4>::Run via a member pointer call
-TEST(SequentialPipelineTest, RunImageToVec3N4CoversRun) {
-    SequentialPipeline<Image, Vec3, 4> pipeline;
-    ImageToPointsStage stage1;
-    PointsToVec3Stage stage2;
-
-    pipeline.AddStage(stage1).Complete(stage2);
-
-    Image input{};
-    auto runPtr = &SequentialPipeline<Image, Vec3, 4>::Run;
-    Vec3 result = (pipeline.*runPtr)(input);
-
-    ASSERT_EQ(result.x, 5);
-    ASSERT_EQ(result.y, 6);
-    ASSERT_EQ(result.z, 7);
-}
-
-// Also verify the alias instantiation uses the same Run logic
-TEST(SequentialPipelineTest, RunDistancePipelineAliasViaPointer) {
-    DistancePipeline pipeline;
-    ImageToPointsStage stage1;
-    PointsToVec3Stage stage2;
-
-    pipeline.AddStage(stage1).Complete(stage2);
-
-    Image input{};
-    auto runPtr = &DistancePipeline::Run;
-    PositionVector result = (pipeline.*runPtr)(input);
-
-    ASSERT_EQ(result.x, 5);
-    ASSERT_EQ(result.y, 6);
-    ASSERT_EQ(result.z, 7);
+    ASSERT_THAT(actual, testing::UnorderedElementsAreArray(matchers));
 }
